+++ RealmServer/Handlers/OnGmTicketGetTicket.cs
@@ -8,7 +8,6 @@ namespace RealmServer.Handlers
         public static void Handler(RealmServerSession session, byte[] data)
         {
             //session.SendPacket(new SMSG_GMTICKET_GETTICKET(TicketInfoResponse.NoTicket));
-
             session.SendPacket(new SMSG_GMTICKET_GETTICKET(TicketInfoResponse.Pending, "TEXTO"));
             session.SendPacket(new SMSG_QUERY_TIME_RESPONSE());
         }