+++ RealmServer/mainProgram.cs
@@ -13,6 +13,7 @@ using Common.Helpers;
 using MongoDB.Driver;
 using RealmServer.Handlers;
 using RealmServer.PacketReader;
+using RealmServer.World.Managers;
 
 namespace RealmServer
 {
@@ -76,10 +77,12 @@ namespace RealmServer
             Log.Print(LogType.RealmServer, $"Running on .NET Framework Version {Environment.Version}");
 
             ConfigFile();
-
+            //
             XmlReader.Boot();
             //
             DbcInit();
+            //
+            PlayerManager.Boot();
             //
             Initalizing();
 
@@ -481,7 +484,6 @@ namespace RealmServer
             if (item == null)
                 return;
 
-            
             session.SendPacket(new SMSG_ITEM_QUERY_SINGLE_RESPONSE(item));
         }
 
@@ -540,14 +542,113 @@ Commands:
     {
         public SMSG_ITEM_QUERY_SINGLE_RESPONSE(Items item) : base(RealmEnums.SMSG_ITEM_QUERY_SINGLE_RESPONSE)
         {
-            Write((UInt32) item.Entry); // 32
-            Write((UInt32) item.Class); //32
-            Write((UInt32) item.SubClass); // 32
+            try
+            {
+                // Item ID
+                Write(item.Entry);
+                // Item Class
+                Write(item.Class);
+                // Item SubCLass
+                Write(item.SubClass);
+                // Item Name
+                WriteCString(item.Name);
+                // Item Name
+                WriteCString(null);
+                // Item Name
+                WriteCString(null);
+                // Item Name
+                WriteCString(null);
 
-            WriteCString(item.Name); // string
-            WriteCString(string.Empty);
-            WriteCString(string.Empty);
-            WriteCString(string.Empty);
+                // Model Id
+                Write(item.DisplayId);
+                // Quality Id
+                Write(item.Quality);
+
+                // Flags
+                Write(item.Flags);
+                // Buy Price
+                Write(12); // BuyCount??
+                // Sell Price
+                Write(34);
+                // Inventory Type
+                Write((int) item.InventoryType);
+
+                // Req Class
+                Write(item.AllowableClass);
+                // Req Race
+                Write(item.AllowableRace);
+
+                // Level
+                Write(item.ItemLevel);
+                // Req Level
+                Write(0);
+
+                // Skill Req
+                Write(0);
+                // Skill Level Req
+                Write(0);
+
+                // Item Unique
+                Write(0);
+                // Max Stack
+                Write(item.MaxCount);
+                // Container Slots
+                Write(0);
+
+                for (int a = 0; a < 10; a++)
+                {
+                    Write(0); // Item Attributes
+                    Write(0); // Item Attributes
+                }
+
+                for (int b = 0; b < 5; b++)
+                {
+                    Write(0); // Minimum Damage
+                    Write(0); // Maximum Damage
+                    Write(0); // Damage Type
+                }
+
+                Write(0); // Physical (Bonus) "Armor"
+                Write(0); // Holy (BONUS)
+                Write(0); // Fire (BONUS)
+                Write(0); // Nature (BONUS)
+                Write(0); // Frost (BONUS)
+                Write(0); // Shadow (BONUS)
+                Write(0); // Item Attack Speed.
+                Write(0); // Ammo Type
+                Write(0); // Max Durability
+
+                for (int c = 0; c < 5; c++) // Spell Info
+                {
+                    Write(0); // Category);
+                    Write(0); // CategoryCoolDown);
+                    Write(0); // Charges);
+                    Write(-1); // Cooldown);
+                    Write(0); // ID);
+                    Write(-1); // Trigger);
+                }
+
+                Write(item.Bonding); // Bonding
+
+                WriteCString(item.Description); // Item Description
+
+                Write(0); // PageText
+                Write(0); // Language ID
+                Write(0); // Page Material;
+                Write(0); // Start Quest
+                Write(0); // Lock
+                Write(0); // Material
+                Write(0); // SheAtTheType;
+                Write(0); // Unknown
+                Write(0); // Unknown 
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e.Message);
+                var trace = new StackTrace(e, true);
+                Log.Print(LogType.Error,
+                    $"{e.Message}: {e.Source}\n{trace.GetFrame(trace.FrameCount - 1).GetFileName()}:{trace.GetFrame(trace.FrameCount - 1).GetFileLineNumber()}");
+            }
         }
     }
 }